+++ R/ksTest.R
@@ -25,7 +25,9 @@
 #' ## see ks.test for other examples
 #' x <- rnorm(50)
 #' y <- runif(30)
-#' df <- data.frame(dat=c(x,y),grp=factor(rep(c("x","y"),c(50,30))))
+#' df <- data.frame(dat=c(x,y),
+#'                  grp=factor(rep(c("x","y"),c(50,30))),
+#'                  stringsAsFactors=FALSE)
 #' 
 #' ## one-sample (from ks.test) still works
 #' ksTest(x+2, "pgamma", 3, 2)