+++ tests/testthat/testthat_dunnTest.R
@@ -55,15 +55,15 @@ test_that("dunnTest() output",{
 })
 
 test_that("Check dunnTest() output labels",{
-  ## This issue stems from a user complaint that labels for factors that
-  ## don't exist (as would happen from using subset() or filter() screw
-  ## up the results.
+  ## This issue stems from a user complaint that labels for factors that don't
+  ## exist (as would happen with subset() or filter()) mess up the results.
   ponds3 <- subset(ponds,pond!=3)
   tmp2 <- dunnTest(pH~fpond,data=subset(ponds,pond!=3))
-  expect_false(any(grepl("3",levels(tmp2$res$Comparison))))
-  expect_true(any(grepl("1",levels(tmp2$res$Comparison))))
-  expect_true(any(grepl("2",levels(tmp2$res$Comparison))))
-  expect_true(any(grepl("4",levels(tmp2$res$Comparison))))
+  lvls <- as.character(unique(tmp2$res$Comparison))
+  expect_false(any(grepl("3",lvls)))
+  expect_true(any(grepl("1",lvls)))
+  expect_true(any(grepl("2",lvls)))
+  expect_true(any(grepl("4",lvls)))
 })
 
 ## Validate Results ----